8. Project - Image Recognition Restful API/web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
numpy
tensorflow
requests
keras
pillow